+++ src/main/java/net/silentchaos512/scalinghealth/ScalingHealth.java
@@ -36,7 +36,7 @@ public class ScalingHealth implements IModBase {
     public static final String VERSION = "1.3.33";
     public static final String VERSION_SILENTLIB = "3.0.9";
     public static final int BUILD_NUM = 0;
-    public static final String DEPENDENCIES = "required-after:silentlib@[" + VERSION_SILENTLIB + ",);after:morpheus";
+    public static final String DEPENDENCIES = "required-after:forge@[14.23.5.2779,);required-after:silentlib@[" + VERSION_SILENTLIB + ",);after:morpheus";
     public static final String RESOURCE_PREFIX = MOD_ID_LOWER + ":";
 
     public static final String GAME_RULE_DIFFICULTY = "ScalingHealthDifficulty";