+++ CHANGELOG.md
@@ -3,6 +3,9 @@
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.3.34] - 2019-04-07
+Set minimum Forge version to 2779 to stop the crashes
+
 ## [1.3.33] - 2019-02-17
 ### Added
 - Config to disable heart color looping [#155]